+++ electron/main.ts
@@ -1,6 +1,6 @@
-import { app, BrowserWindow, ipcMain } from 'electron';
+import { app, BrowserWindow, ipcMain, ipcRenderer } from 'electron';
 import { runningNotification, stillRunningNotification } from './notifications/notifcation';
-import { HomeWidown } from './ui/windows/home';
+import { HomeWindow } from './ui/windows/home';
 import { Tray } from './ui/tray';
 import { Shortcuts } from './helpers/shortcuts';
 import { store } from './store/config';
@@ -9,7 +9,9 @@ import { autoUpdater } from "electron-updater"
 import path from 'path';
 import fs from 'fs';
 import fsp from 'fs/promises';
-import { spawn,  } from 'child_process';
+import { spawn, } from 'child_process';
+import { PromptsWindow } from './ui/windows/prompts';
+import { UserSettingsSchema } from './store/schemas/userSettings'
 
 const recordingsDir = path.join(app.getPath('documents'), app.getName(), 'recordings');
 
@@ -26,8 +28,9 @@ if (!gotTheLock) {
 } else {
 
   const App = () => {
-    HomeWidown.getInstance()
     SettingsWindow.getInstance().bw.hide()
+    PromptsWindow.getInstance().bw.hide()
+
     Tray.getInstance()
   }
 
@@ -40,7 +43,7 @@ if (!gotTheLock) {
 
   app.on('activate', () => {
     if (BrowserWindow.getAllWindows().length === 0) {
-      HomeWidown.getInstance()
+      HomeWindow.getInstance()
     }
   });
 
@@ -54,18 +57,25 @@ if (!gotTheLock) {
       ipcMain.on('electron-store-set', async (e: Electron.IpcMainEvent, key: string, value: unknown) => {
         store.set(key, value);
       });
+      ipcMain.on('electron-store-del', async <Key extends keyof typeof UserSettingsSchema>(e: Electron.IpcMainEvent, key: Key) => {
+        store.delete(key)
+        console.log("delete: " + key)
+      });
 
       ipcMain.on('electron-store-open-editor', async () => {
         store.openInEditor()
       });
 
-      ipcMain.on('open-window', async (e: Electron.IpcMainEvent, windowName: string) => {
+      ipcMain.on('open-window', async (e: Electron.IpcMainEvent, windowName: string, ...args: unknown[]) => {
         switch (windowName) {
           case "home":
-            HomeWidown.getInstance().bw.show()
+            HomeWindow.getInstance().bw.show()
             break;
           case "settings":
             SettingsWindow.getInstance().bw.show()
+            break;
+          case "prompts":
+            PromptsWindow.getInstance().bw.show()
             break;
         }
       });
@@ -74,10 +84,13 @@ if (!gotTheLock) {
 
         switch (windowName) {
           case "home":
-            HomeWidown.getInstance().bw.hide()
+            HomeWindow.getInstance().bw.hide()
             break;
           case "settings":
             SettingsWindow.getInstance().bw.hide()
+            break;
+          case "prompts":
+            PromptsWindow.getInstance().bw.hide()
             break;
         }
       })
@@ -86,7 +99,7 @@ if (!gotTheLock) {
 
         switch (windowName) {
           case "home":
-            HomeWidown.getInstance().bw.minimize()
+            HomeWindow.getInstance().bw.minimize()
             break;
           case "settings":
             SettingsWindow.getInstance().bw.minimize()
@@ -141,6 +154,10 @@ if (!gotTheLock) {
           console.log(`transcript: child process exited with code ${code}`);
         });
       });
+
+      ipcMain.on('dispatch-sync-config', async (e: Electron.IpcMainEvent, key: string, value: unknown) => {
+        HomeWindow.getInstance().bw.webContents.send('sync-config', { key, value });
+      })
 
     })
     .then(() => {