+++ electron/store/schemas/userSettings.ts
@@ -1,10 +1,14 @@
 import { JSONSchemaType } from 'json-schema-typed'
 
-export const userSettingsSchema = {
+export const UserSettingsSchema = {
 	general: {
 		type: JSONSchemaType.Object,
 		properties: {
 			theme: { type: JSONSchemaType.String, },
+			backgroundOpacity: {
+				type: JSONSchemaType.Boolean,
+				default: false
+			},
 			language: { type: JSONSchemaType.String },
 			chatMode: {
 				type: JSONSchemaType.Object,
@@ -67,6 +71,20 @@ export const userSettingsSchema = {
 		default: {
 			current: 'gemini',
 		}
+	},
+	prompts: {
+		type: JSONSchemaType.Object,
+		additionalProperties: {
+			type: JSONSchemaType.Object,
+			properties: {
+				title: { type: JSONSchemaType.String },
+				content: { type: JSONSchemaType.String },
+				selected: { type: JSONSchemaType.Boolean },
+				default: { type: JSONSchemaType.Boolean }
+			},
+			// required: ['title', 'content']
+		},
+		default: {}
 	},
 	tabs: {
 		type: JSONSchemaType.Object,