+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Unreleased
 
+- chore: add background opacity option
+- fix: MaxListenersExceededWarning
+- fix: UnhandledPromiseRejectionWarning
+
 ## v1.5.3 - 2025-09-14
 
 - feat: add bump-my-version